🚚 Swap /store/featured & /store/index page
